# packages/servers/analytics-server/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nodash Analytics Server",
    "description": "Analytics data collection and processing server",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.nodash.ai/analytics",
      "description": "Production analytics server"
    }
  ],
  "paths": {
    "/events": {
      "post": {
        "operationId": "trackEvent",
        "summary": "Track an analytics event",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event tracked successfully"
          }
        }
      }
    },
    "/events/batch": {
      "post": {
        "operationId": "trackEventsBatch",
        "summary": "Track multiple analytics events",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch processed successfully"
          }
        }
      }
    },
    "/health": {
      "get": {
        "operationId": "getHealthStatus",
        "summary": "Get server health status",
        "responses": {
          "200": {
            "description": "Server is healthy"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EventRequest": {
        "type": "object",
        "required": ["eventName", "projectId"],
        "properties": {
          "eventName": { "type": "string" },
          "projectId": { "type": "string" },
          "properties": { "type": "object" }
        }
      },
      "BatchEventRequest": {
        "type": "object",
        "required": ["events"],
        "properties": {
          "events": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/EventRequest" }
          }
        }
      }
    }
  }
}